+++ src/app/guards/signed-in/signed-in.guard.spec.ts
@@ -1,22 +1,14 @@
 import { TestBed, fakeAsync } from '@angular/core/testing';
-import {
-  ActivatedRouteSnapshot,
-  CanActivateFn,
-  Router,
-  RouterStateSnapshot,
-  UrlSegment,
-  UrlSegmentGroup,
-  UrlTree,
-} from '@angular/router';
+import { CanActivateFn, Router, UrlSegment, UrlSegmentGroup, UrlTree } from '@angular/router';
 
-import { signinedInGuard } from './signined-in.guard';
+import { signedInGuard } from './signed-in.guard';
 import { AuthService } from '@services';
-import { throwError, Observable, of } from 'rxjs';
-import { runAuthGuardWithContext } from '../utils';
+import { throwError, of } from 'rxjs';
+import { getGuardWithDummyUrl, runAuthGuardWithContext } from '../utils';
 
-describe('signinedInGuard', () => {
+describe('signedInGuard', () => {
   const executeGuard: CanActivateFn = (...guardParameters) =>
-    TestBed.runInInjectionContext(() => signinedInGuard(...guardParameters));
+    TestBed.runInInjectionContext(() => signedInGuard(...guardParameters));
 
   let mockAuthService: jasmine.SpyObj<AuthService>;
   let mockRouter: jasmine.SpyObj<Router>;
@@ -25,8 +17,8 @@ describe('signinedInGuard', () => {
   const expectedUrl = '/';
 
   beforeEach(() => {
-    mockAuthService = jasmine.createSpyObj(signinedInGuard, ['checkAuthenticated']);
-    mockRouter = jasmine.createSpyObj(signinedInGuard, ['navigate', 'createUrlTree', 'parseUrl']);
+    mockAuthService = jasmine.createSpyObj(signedInGuard, ['checkAuthenticated']);
+    mockRouter = jasmine.createSpyObj(signedInGuard, ['navigate', 'createUrlTree', 'parseUrl']);
     mockRouter.parseUrl.and.callFake((url: string) => {
       const urlTree = new UrlTree();
       const urlSegment = new UrlSegment(url, {});
@@ -54,25 +46,16 @@ describe('signinedInGuard', () => {
 
   it('should return false if the user is logged in and redirect to the default route', fakeAsync(async () => {
     mockIsLoggedInFalse();
-    const authenticated = await runAuthGuardWithContext(getSignedInGuardWithDummyUrl(urlPath));
+    const authenticated = await runAuthGuardWithContext(getGuardWithDummyUrl(signedInGuard, urlPath));
     expect(mockRouter.navigate).toHaveBeenCalledOnceWith([expectedUrl]);
     expect(authenticated).toBeFalsy();
   }));
 
   it('should allow access to login if catches an error ', fakeAsync(async () => {
     mockAuthService.checkAuthenticated.and.returnValue(throwError(() => 'Authentication error'));
-    const authenticated = await runAuthGuardWithContext(getSignedInGuardWithDummyUrl(urlPath));
+    const authenticated = await runAuthGuardWithContext(getGuardWithDummyUrl(signedInGuard, urlPath));
     expect(authenticated).toBeTruthy();
   }));
-
-  function getSignedInGuardWithDummyUrl(
-    urlPath: string,
-  ): () => boolean | UrlTree | Promise<boolean | UrlTree> | Observable<boolean | UrlTree> {
-    const dummyRoute = new ActivatedRouteSnapshot();
-    dummyRoute.url = [new UrlSegment(urlPath, {})];
-    const dummyState: RouterStateSnapshot = { url: urlPath, root: new ActivatedRouteSnapshot() };
-    return () => signinedInGuard(dummyRoute, dummyState);
-  }
 
   const mockIsLoggedInFalse = () => {
     mockAuthService.checkAuthenticated.and.returnValue(of(false));
