+++ src/app/guards/signed-in/signed-in.guard.ts
@@ -3,7 +3,7 @@ import { CanActivateFn, Router } from '@angular/router';
 import { AuthService } from '@services';
 import { map, catchError, of } from 'rxjs';
 
-export const signinedInGuard: CanActivateFn = (route, state) => {
+export const signedInGuard: CanActivateFn = () => {
   const authService: AuthService = inject(AuthService);
   const router = inject(Router);
   return authService.checkAuthenticated().pipe(
