+++ src/app/guards/utils/index.ts
@@ -1,7 +1,19 @@
-import { UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot, UrlSegment, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 
 import { TestBed } from '@angular/core/testing';
+
+import { authGuard, signedInGuard } from '@guards';
+
+export function getGuardWithDummyUrl(
+  guard: typeof authGuard | typeof signedInGuard,
+  urlPath: string,
+): () => boolean | UrlTree | Promise<boolean | UrlTree> | Observable<boolean | UrlTree> {
+  const dummyRoute = new ActivatedRouteSnapshot();
+  dummyRoute.url = [new UrlSegment(urlPath, {})];
+  const dummyState: RouterStateSnapshot = { url: urlPath, root: new ActivatedRouteSnapshot() };
+  return () => guard(dummyRoute, dummyState);
+}
 
 export async function runAuthGuardWithContext(
   authGuard: () => boolean | UrlTree | Promise<boolean | UrlTree> | Observable<boolean | UrlTree>,
